chore: update hcloud API SDK to v2
There are no functional changes, but SDK got updated to handle int ->
int64 changes. v1 version is only supported to Sep 2023.

See https://github.com/hetznercloud/hcloud-go#support

+++ pkg/machinery/resources/network/operator_spec.go
@@ -85,8 +85,8 @@ type VIPEquinixMetalSpec struct {
 //
 //gotagsrewrite:gen
 type VIPHCloudSpec struct {
-	DeviceID  int    `yaml:"deviceID" protobuf:"1"`
-	NetworkID int    `yaml:"networkID" protobuf:"2"`
+	DeviceID  int64  `yaml:"deviceID" protobuf:"1"`
+	NetworkID int64  `yaml:"networkID" protobuf:"2"`
 	APIToken  string `yaml:"apiToken" protobuf:"3"`
 }
